move body templates from table to separate compopnents
