+++ src/components/body-templates/rating.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import { Rating } from 'primereact/rating';
+
+const ratingBodyTemplate = (item) => {
+    return <Rating value={item.rating} readOnly stars={10} />;
+};
+
+export default ratingBodyTemplate;
