+++ src/components/table.js
@@ -1,14 +1,14 @@
 import React from 'react';
-import { useSelector, useDispatch } from 'react-redux';
 import { DataTable } from 'primereact/datatable';
 import { Column } from 'primereact/column';
 import { InputText } from 'primereact/inputtext';
 import { Dropdown } from 'primereact/dropdown';
-import { Rating } from 'primereact/rating';
-import { Button } from 'primereact/button';
 import { types, statuses } from '../../constants';
-import { setIsSidebarVisible } from '../redux/actions/action-creators';
-import { Tag } from 'primereact/tag';
+import statusBodyTemplate from './body-templates/status';
+import typeBodyTemplate from './body-templates/type';
+import ratingBodyTemplate from './body-templates/rating';
+import showSidebarButtonBodyTemplate from './body-templates/show-sidebar-button';
+import { Rating } from 'primereact/rating';
 
 const Table = ({
     items,
@@ -18,8 +18,6 @@ const Table = ({
     onShowDeletedItemsClick,
     handleDeleteItems */
 }) => {
-    const dispatch = useDispatch();
-
     // todo: move editors to ./editors/...
     const textEditor = (options) => {
         return (
@@ -54,11 +52,6 @@ const Table = ({
         );
     };
 
-    // todo: move body templates to ./body-templates/...
-    const ratingBodyTemplate = (item) => {
-        return <Rating value={item.rating} readOnly stars={10} />;
-    };
-
     const ratingEditor = (options) => {
         return (
             <Rating
@@ -76,82 +69,6 @@ const Table = ({
         if (wasChanged) {
             handleSubmit(rowData, newRowData);
         }
-    };
-
-    const handleShow = (rowData) => {
-        console.log('bb ~ handleShow ~ rowData:', rowData);
-        dispatch(setIsSidebarVisible(true, rowData));
-    };
-    const buttonBodyTemplate = (rowData) => {
-        return (
-            <Button
-                type='button'
-                icon='pi pi-info-circle'
-                className='p-button-rounded p-button-warning'
-                onClick={() => handleShow(rowData)}
-            />
-        );
-    };
-
-    const typeBodyTemplate = (item) => {
-        const itemType = item && item.type ? item.type : '';
-        let iconCode = '';
-        switch (itemType) {
-            case 'Book':
-                iconCode = 'pi pi-book';
-                break;
-            case 'Movie':
-                iconCode = 'pi pi-video';
-                break;
-            case 'TV':
-                iconCode = 'pi pi-desktop';
-                break;
-            case 'Video Game':
-                iconCode = 'pi pi-discord';
-                break;
-            default:
-                iconCode = 'pi pi-question';
-        }
-
-        return (
-            <span className={`type-badge type-${itemType.toLowerCase()}`}>
-                <i className={`pi ${iconCode}`}></i>
-                {itemType}
-            </span>
-        );
-    };
-
-    const statusBodyTemplate = (item) => {
-        const status = item && item.status ? item.status : '';
-        let color = '';
-        let iconCode = '';
-        switch (status) {
-            case 'Now':
-                color = 'warning';
-                iconCode = 'pi pi-play';
-                break;
-            case 'Past':
-                color = 'success';
-                iconCode = 'pi pi-check';
-                break;
-            case 'Future':
-                color = 'info';
-                iconCode = 'pi pi-calendar';
-                break;
-            case 'On Hold':
-                color = 'info';
-                iconCode = 'pi pi-pause';
-                break;
-            case 'Dropped':
-                color = 'danger';
-                iconCode = 'pi pi-times';
-                break;
-            default:
-                color = 'secondary';
-                iconCode = 'pi pi-question';
-        }
-
-        return <Tag value={status} severity={color} icon={iconCode} />;
     };
 
     return (
@@ -190,7 +107,7 @@ const Table = ({
                 body={ratingBodyTemplate}
                 editor={(options) => ratingEditor(options)}
                 onCellEditComplete={onCellEditComplete}></Column>
-            <Column header='Details' body={buttonBodyTemplate}></Column>
+            <Column header='Details' body={showSidebarButtonBodyTemplate}></Column>
         </DataTable>
     );
 };
